+++ Editor/OculusMetadata.cs
@@ -40,13 +40,8 @@ namespace Unity.XR.Oculus.Editor
             var settings = obj as OculusSettings;
             if (settings != null)
             {
-#if UNITY_2021_2_OR_NEWER
                 settings.m_StereoRenderingModeDesktop = OculusSettings.StereoRenderingModeDesktop.SinglePassInstanced;
                 settings.m_StereoRenderingModeAndroid = OculusSettings.StereoRenderingModeAndroid.Multiview;
-#else
-                settings.m_StereoRenderingModeDesktop = OculusSettings.StereoRenderingModeDesktop.MultiPass;
-                settings.m_StereoRenderingModeAndroid = OculusSettings.StereoRenderingModeAndroid.MultiPass;
-#endif
                 settings.SharedDepthBuffer = true;
                 settings.DashSupport = true;
                 settings.LowOverheadMode = false;
@@ -61,6 +56,7 @@ namespace Unity.XR.Oculus.Editor
                 settings.TargetQuestPro = false;
                 settings.TargetQuest3 = false;
                 settings.DepthSubmission = false;
+                settings.UseStickControlThumbsticks = false;
 
                 return true;
             }