+++ Runtime/OculusLoader.cs
@@ -367,52 +367,6 @@ namespace Unity.XR.Oculus
 #endif
 
 #if UNITY_EDITOR && XR_MGMT_GTE_320
-#if !UNITY_2021_2_OR_NEWER
-        private void RemoveVulkanFromAndroidGraphicsAPIs()
-        {
-
-            // don't need to do anything if auto apis is selected
-            if (PlayerSettings.GetUseDefaultGraphicsAPIs(BuildTarget.Android))
-                return;
-
-            GraphicsDeviceType[] oldApis = PlayerSettings.GetGraphicsAPIs(BuildTarget.Android);
-            List<GraphicsDeviceType> newApisList = new List<GraphicsDeviceType>();
-            bool vulkanRemoved = false;
-
-            // copy all entries except vulkan
-            foreach (GraphicsDeviceType dev in oldApis)
-            {
-                if (dev == GraphicsDeviceType.Vulkan)
-                {
-                    vulkanRemoved = true;
-                    continue;
-                }
-
-                newApisList.Add(dev);
-            }
-
-            // if we didn't remove Vulkan from the list, no need to do any further processing
-            if (vulkanRemoved == false)
-                return;
-
-            if (newApisList.Count <= 0)
-            {
-                newApisList.Add(GraphicsDeviceType.OpenGLES3);
-                Debug.LogWarning(
-                    "Vulkan is currently experimental on Oculus Quest. It has been removed from your list of Android graphics APIs and replaced with OpenGLES3.\n" +
-                    "If you would like to use experimental Quest Vulkan support, you can add it back into the list of graphics APIs in the Player settings.");
-            }
-            else
-            {
-                Debug.LogWarning(
-                    "Vulkan is currently experimental on Oculus Quest. It has been removed from your list of Android graphics APIs.\n" +
-                    "If you would like to use experimental Quest Vulkan support, you can add it back into the list of graphics APIs in the Player settings.");
-            }
-
-            PlayerSettings.SetGraphicsAPIs(BuildTarget.Android, newApisList.ToArray());
-        }
-#endif // !UNITY_2021_2_OR_NEWER
-
         private void SetAndroidMinSdkVersion()
         {
             if (PlayerSettings.Android.minSdkVersion < AndroidSdkVersions.AndroidApiLevel23)
@@ -426,9 +380,6 @@ namespace Unity.XR.Oculus
         {
             if (buildTargetGroup == BuildTargetGroup.Android)
             {
-#if !UNITY_2021_2_OR_NEWER
-                RemoveVulkanFromAndroidGraphicsAPIs();
-#endif
                 SetAndroidMinSdkVersion();
             }
         }